fix: cast cache result to int in getUnreadNotificationsCountCached()
Fixes TypeError in production where Cache::remember() returns string
instead of int, causing 'Return value must be of type int, string
returned' error.

+++ app/Models/User.php
@@ -218,7 +218,7 @@ final class User extends Authenticatable implements MustVerifyEmail
 
     public function getUnreadNotificationsCountCached(): int
     {
-        return Cache::remember(
+        return (int) Cache::remember(
             "user:{$this->id}:unread_notifications_count",
             now()->addSeconds(30),
             fn () => $this->unreadNotifications()->count()
